refactored all the server files into another file called server and use the index file to start the server
