+++ lib/handlers.js
@@ -4,7 +4,7 @@
  */
 var helpers = require('./helpers');
 var _data = require('./data');
-var config = require('./../config');
+var config = require('./config');
 var handlers = {}; //container for handlers
 
 //users handler\
