+++ lib/helpers.js
@@ -5,7 +5,7 @@
 
 //dependencies
 var crypto = require('crypto');
-var config = require('./../config');
+var config = require('./config');
 var https = require('https');
 var querystring = require('querystring');
